# api-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Apartment Management API",
        "description": "API for managing apartments in a tourist management system.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "paths": {
...
        "/apartment/{idApartment}": {
            "get": {
                "summary": "Get details of a specific apartment",
                "tags": [
                    "Public"
                ],
                "parameters": [
                    {
                        "name": "idApartment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieves the details of the apartment"
                    },
                    "404": {
                        "description": "Apartment not found"
                    }
                }
            }
        },
...
    }
}